# public/swagger-config.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Streaming API",
        "version": "1.0.0",
        "description": "API documentation for the Streaming application",
        "contact": {
            "name": "Streaming API",
            "email": "support@streaming.araneum.ba"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Genre": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tmdb_id": {
                        "type": "string",
                        "example": "28"
                    },
                    "name": {
                        "type": "string",
                        "example": "Action"
                    },
                    "type": {
                        "type": "string",
                        "example": "movie"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user details",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get list of movies",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Movie"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/trending": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get trending movies",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of trending movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Movie"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/upcoming": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get upcoming movies",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of upcoming movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Movie"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{slug}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get movie details by slug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "description": "Movie slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Movie"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "sources": {
                                                    "type": "object",
                                                    "properties": {
                                                        "vidsrc": {
                                                            "type": "string",
                                                            "example": "https://vidsrc.to/embed/movie/299534"
                                                        },
                                                        "2embed": {
                                                            "type": "string",
                                                            "example": "https://2embed.to/embed/movie/299534"
                                                        }
                                                    }
                                                },
                                                "credits": {
                                                    "type": "object"
                                                },
                                                "videos": {
                                                    "type": "object"
                                                },
                                                "similar": {
                                                    "type": "object"
                                                },
                                                "recommendations": {
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            }
        },
        "/tv-shows": {
            "get": {
                "tags": [
                    "TV Shows"
                ],
                "summary": "Get list of TV shows",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of TV shows",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TvShow"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tv-shows/trending": {
            "get": {
                "tags": [
                    "TV Shows"
                ],
                "summary": "Get trending TV shows",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of trending TV shows",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TvShow"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tv-shows/{slug}": {
            "get": {
                "tags": [
                    "TV Shows"
                ],
                "summary": "Get TV show details by slug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "description": "TV show slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TV show details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TvShow"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "seasons": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "credits": {
                                                    "type": "object"
                                                },
                                                "videos": {
                                                    "type": "object"
                                                },
                                                "similar": {
                                                    "type": "object"
                                                },
                                                "recommendations": {
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TV show not found"
                    }
                }
            }
        },
        "/tv-shows/{slug}/season/{seasonNumber}": {
            "get": {
                "tags": [
                    "TV Shows"
                ],
                "summary": "Get season details",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "description": "TV show slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seasonNumber",
                        "in": "path",
                        "required": true,
                        "description": "Season number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Season details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "tv_show": {
                                            "$ref": "#/components/schemas/TvShow"
                                        },
                                        "season": {
                                            "type": "object"
                                        },
                                        "episodes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Episode"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TV show or season not found"
                    }
                }
            }
        },
        "/tv-shows/{slug}/season/{seasonNumber}/episode/{episodeNumber}": {
            "get": {
                "tags": [
                    "TV Shows"
                ],
                "summary": "Get episode details",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "description": "TV show slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seasonNumber",
                        "in": "path",
                        "required": true,
                        "description": "Season number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "episodeNumber",
                        "in": "path",
                        "required": true,
                        "description": "Episode number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "tv_show": {
                                            "$ref": "#/components/schemas/TvShow"
                                        },
                                        "episode": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/Episode"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "sources": {
                                                            "type": "object",
                                                            "properties": {
                                                                "vidsrc": {
                                                                    "type": "string",
                                                                    "example": "https://vidsrc.to/embed/tv/1399/1/1"
                                                                },
                                                                "2embed": {
                                                                    "type": "string",
                                                                    "example": "https://2embed.to/embed/series/1399/1/1"
                                                                }
                                                            }
                                                        },
                                                        "credits": {
                                                            "type": "object"
                                                        },
                                                        "images": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TV show, season, or episode not found"
                    }
                }
            }
        },
...
        "/genres/{type}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get genres by type",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "description": "Genre type (movie or tv)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "movie",
                                "tv"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of genres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/genre/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get genre details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Genre ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Genre details with related movies and TV shows",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "genre": {
                                            "$ref": "#/components/schemas/Genre"
                                        },
                                        "movies": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/Pagination"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Movie"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "tv_shows": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/Pagination"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/TvShow"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Genre not found"
                    }
                }
            }
        },
        "/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search for movies and TV shows",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "description": "Search query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "movies": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            }
                                        },
                                        "tv_shows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TvShow"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/watchlist": {
            "get": {
                "tags": [
                    "Watchlist"
                ],
                "summary": "Get user's watchlist",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's watchlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "movies": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            }
                                        },
                                        "tv_shows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TvShow"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}